requirements.txt
adbc_driver_manager==1.3.0
adbc_driver_postgresql==1.3.0
adbc_driver_sqlite==1.3.0
AppKit==0.2.8
argcomplete==3.5.2
asgiref==3.8.1
atomicwrites==1.4.1
beautifulsoup4==4.12.3
boto3==1.35.77
botocore==1.35.77
brotli==1.1.0
brotlicffi==1.1.0.0
celery==5.4.0
chardet==5.2.0
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.9
ConfigParser==7.1.0
cryptography==44.0.0
cycler==0.11.0
Cython==3.0.11
dash_leaflet==1.0.15
dash_testing_stub==0.0.2
dask==2024.12.0
diskcache==5.6.3
dl==0.1.0
docutils==0.21.2
eventlet==0.38.0
fast_dash==0.2.9
fastparquet==2024.11.0
fire==0.7.0
flask_compress==1.17
Foundation==0.1.0a0.dev1
fsspec==2023.12.1
gevent==24.11.1
greenlet==3.1.1
h2==4.1.0
HTMLParser==0.0.2
hypothesis==6.122.3
importlib_resources==6.4.5
ini2toml==0.15
inotify==0.2.10
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.0.2
jnius==1.1.0
js==1.0
keyring==25.5.0
lxml==4.9.1
matplotlib==3.6.0
mtrand==0.1
multiprocess==0.70.17
mypy==1.13.0
numarray==1.5.1
Numeric==24.2
numexpr==2.10.2
odfpy==1.4.1
openpyxl==3.1.5
ordereddict==1.1
orjson==3.10.12
PasteScript==3.6.0
percy==2.0.2
pexpect==4.8.0
pickle5==0.0.12
Pillow==11.0.0
pkg1==0.0.3
protobuf==5.29.1
psutil==5.9.2
PyInstaller==6.11.1
pyodide==0.0.2
pyOpenSSL==24.3.0
PyQt4==4.11.4
PyQt5==5.15.11
python-dotenv==1.0.1
python_calamine==0.3.1
pyxlsb==1.0.10
QtPy==2.2.0
railroad==0.5.0
redis==5.2.1
ruamel.yaml.clib==0.2.12
s3fs==2024.10.0
scikit_learn==1.6.0
scipy==1.14.1
scipy_doctest==1.5.1
selenium==4.27.1
setproctitle==1.3.4
sets==0.3.2
simplejson==3.19.3
skimage==0.0
sniffio==1.3.1
Sphinx==8.1.3
SQLAlchemy==2.0.36
statsmodels==0.14.4
tables==3.10.1
thread==2.0.5
threadpoolctl==3.5.0
tomli_w==1.1.0
tornado==6.2
traitlets==5.14.3
trio==0.27.0
trove_classifiers==2024.10.21.16
Twisted==24.11.0
ukkonen==1.0.1
urllib3_secure_extra==0.1.0
vincent==0.4.4
watchdog==6.0.0
xarray==2024.11.0
xlrd==2.0.1
XlsxWriter==3.2.0
xmlrpclib==1.0.1
xx==3.3.2
zope==5.11.1
zstandard==0.23.0
gunicorn
pandas
dash-tools
